Move batch scan logic to OneScanCovidRDTBarcodeFactory
